avoid different sign comparison

+++ emscripten_audio.cpp
@@ -101,7 +101,7 @@ emscripten_audio::emscripten_audio(construction_options &&options)
               } else {                                                          // if no output function is provided, output silence to avoid generating noise
                 for(unsigned int output_index{0}; output_index != num_outputs; ++output_index) {
                   auto &output{outputs[output_index]};
-                  std::memset(output.data, 0, output.numberOfChannels * output.samplesPerChannel * sizeof(float)); // could use std::fill here but memset is reportedly faster, https://lemire.me/blog/2020/01/20/filling-large-arrays-with-zeroes-quickly-in-c/
+                  std::memset(output.data, 0, static_cast<size_t>(output.numberOfChannels) * static_cast<size_t>(output.samplesPerChannel) * sizeof(float)); // could use std::fill here but memset is reportedly faster, https://lemire.me/blog/2020/01/20/filling-large-arrays-with-zeroes-quickly-in-c/
                 }
               }
               return true;                                                      // keep the graph output going
